docs: add attribution for adapted code
Adds links for XORShift32, Mulberry32 and LCG PRNG implementations,
which are in the public domain.

Adds a link for the sinFract() function, which is adapted from The
Art of Code's YouTube video on Value Noise.

Adds MIT license text and link for Quadtree class, which is adapted
from CodingTrain/QuadTree.

Refs: #72, #73