feature!: rename Rectangle class to AABB (fixes #66)
The Rectangle class cannot be rotated. It is used to do collision and
intersection checks. Renaming the class to AABB allows for a rotatable
Rectangle class in the future.

Tests for the class have been renamed and formatted slightly, and moved
to tests/AABB.spec.js.

BREAKING CHANGE: change Rectangle instances to be of class AABB