refactor(PRNG)!: use strategy pattern for prng algorithms
New PRNG algorithms should conform to the PRNGAlgorithm interface.
A PRNG object can now be instantiated directly, using new. Optionally,
it can be configured with an object. The object can have a `seed`,
with a number value, and an `algorithm`, with the name of the chosen
implementation. For now, can be either LCG, XORShift32, or Mulberry32.

BREAKING CHANGE: PRNG implementations should no longer be created
	directly. Instead, create a PRNG object, and optionally give
	a config object to the constructor specifying seed and chosen
	PRNGAlgorithm